+++ src/routes/(docs)/[[slug]]/docs/Callout.svelte
@@ -85,7 +85,9 @@
 <CodeBlock
 	code={`
     <Callout type="info">
-        <div slot="title">This is a title</div>
+        {#snippet title()}
+            <div>This is a title</div>
+        {/snippet}
         This is a callout with a title.
     </Callout>
 `}
@@ -105,7 +107,9 @@
 <CodeBlock
 	code={`
     <Callout type="info">
-        <div slot="icon">👋</div>
+        {#snippet icon()}
+            <div>👋</div>
+        {/snippet}
         This is a callout with an icon.
     </Callout>
 `}
@@ -125,8 +129,12 @@
 <CodeBlock
 	code={`
     <Callout type="danger">
-        <div slot="icon">⛔️</div>
-        <div slot="title">This is a title</div>
+        {#snippet icon()}
+            <div>⛔️</div>
+        {/snippet}
+        {#snippet title()}
+            <div>This is a title</div>
+        {/snippet}
         This is a callout with an icon and a title.
     </Callout>
 `}
@@ -149,8 +157,12 @@
 <CodeBlock
 	code={`
     <Callout type="info">
-        <div slot="icon">👋</div>
-        <div slot="title">hi there</div>
+        {#snippet icon()}
+            <div>👋</div>
+        {/snippet}
+        {#snippet title()}
+            <div>hi there</div>
+        {/snippet}
         This is a callout with an icon and a title.
     </Callout>
 `}