+++ src/routes/(docs)/[[slug]]/docs/Toast/Message1.svelte
@@ -13,7 +13,7 @@
 This is a Svelte Component notification with a long text message and a button to close it.
 
 <div style="margin-top:10px;text-align:center">
-	<Button on:click={() => toastService.close(toast.id)} size="small" style="margin-left: 4px;"
+	<Button onclick={() => toastService.close(toast.id)} size="small" style="margin-left: 4px;"
 		>Close</Button
 	>
 </div>