+++ src/routes/animation/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import ConsoleLoader from '$lib/components/ConsoleLoader/ConsoleLoader.svelte';
+</script>
+
+<ConsoleLoader logo="https://hyvor.com/api/public/logo/core.svg" size={100} />